Disentangle Mixins and Components
